Improve flag setting/clearing trace logs

+++ daemons/controld/controld_utils.h
@@ -62,7 +62,7 @@ void controld_stop_sched_timer(void);
 void controld_free_sched_timer(void);
 void controld_expect_sched_reply(xmlNode *msg);
 
-void fsa_dump_actions(long long action, const char *text);
+void fsa_dump_actions(uint64_t action, const char *text);
 void fsa_dump_inputs(int log_level, const char *text, long long input_register);
 
 gboolean update_dc(xmlNode * msg);
